create LLMEvaluation
