+++ IntelliNode/index.js
@@ -25,11 +25,11 @@ const HuggingWrapper = require('./wrappers/HuggingWrapper');
 const ReplicateWrapper = require('./wrappers/ReplicateWrapper');
 const AWSEndpointWrapper = require('./wrappers/AWSEndpointWrapper');
 // utils
+const { LLMEvaluation } = require('./utils/LLMEvaluation');
 const AudioHelper = require('./utils/AudioHelper');
 const Config2 = require('./utils/Config2');
 const ConnHelper = require('./utils/ConnHelper');
 const MatchHelpers = require('./utils/MatchHelpers');
-const ModelEvaluation = require('./utils/ModelEvaluation');
 const SystemHelper = require('./utils/SystemHelper');
 const Prompt = require('./utils/Prompt');
 const ProxyHelper = require('./utils/ProxyHelper');
@@ -72,5 +72,5 @@ module.exports = {
   AWSEndpointWrapper,
   Prompt,
   LLamaSageInput,
-  ModelEvaluation
+  LLMEvaluation
 };
